Move WritableStreamDefaultWriter to separate file
